+++ lib/multisync/toplevel.rb
@@ -21,16 +21,9 @@ class Multisync::Toplevel
     false
   end
   
-  # The toplevel never defines a check. It is true by definition.
-  # Don't return a message or a cmd eigher.
-  def check_passed?
-    true
-  end
-  def check_message
-    false
-  end
-  def check_cmd
-    false
+  # The toplevel never defines a check.
+  def checks
+    []
   end
   
   # Turn off source and destination checks on toplevel.